попытался

+++ src/libraly/data/GostHashData.kt
@@ -1,3 +1,3 @@
 package libraly.data
 
-class GostHashData(override val m: Byte, val r: Long, val s: Long) : HashDataRule
+class GostHashData(override val m: Long, val r: Long, val s: Long) : HashDataRule
