+++ src/libraly/data/RsaHashData.kt
@@ -1,3 +1,3 @@
 package libraly.data
 
-class RsaHashData(override val m: Byte, val s: Long) : HashDataRule
+class RsaHashData(override val m: Long, val s: Long) : HashDataRule
